Все доделано до рефакторинга

+++ app/src/main/java/com/abdulmanov/MoviCorn/ui/details_movie/DetailsMoviePresenter.kt
@@ -1,17 +1,15 @@
 package com.abdulmanov.MoviCorn.ui.details_movie
-import com.abdulmanov.core.database.db.entity.Movie
+
 import com.abdulmanov.MoviCorn.R
-import com.abdulmanov.core.network.model.Model as NetModel
-import com.abdulmanov.core.database.model.Model as DBModel
-import com.abdulmanov.MoviCorn.model.mappers.movies.MoviesDetailsDTOtoDetailsMovieMapper
-import com.abdulmanov.MoviCorn.model.vo.movie.DetailsMovie
+import com.abdulmanov.domain.models.movies.MovieDetails
+import com.abdulmanov.domain.repositories.MoviesDbRepository
+import com.abdulmanov.domain.repositories.MoviesRepository
 import io.reactivex.disposables.Disposable
 
 
 class DetailsMoviePresenter(
-    private val network: NetModel,
-    private val database:DBModel,
-    private val movieMapper:MoviesDetailsDTOtoDetailsMovieMapper
+    private val network: MoviesRepository,
+    private val database:MoviesDbRepository
 ): DetailsMovieContract.Presenter {
 
     private var view: DetailsMovieContract.View? = null
@@ -26,40 +24,32 @@ class DetailsMoviePresenter(
         requestDisposable?.dispose()
     }
 
-    override fun saveMovieInLibrary(movie: DetailsMovie) {
-        requestDisposable = database.existsFilm(movie.id).subscribe(
+    override fun saveMovieInLibrary(movie: MovieDetails) {
+        requestDisposable = database.existsMovie(movie.id).subscribe(
             {
-                database.deleteMovie(movie.id).subscribe {
-                    view?.showMessage(R.string.details_delete_film)
-                    view?.showSaved(false)
+                if(it.isEmpty()){
+                    database.insertMovie(movie).subscribe{
+                        view?.showMessage(R.string.details_save_film)
+                        view?.showSaved(true)
+                    }
+                }else{
+                    database.deleteMovie(movie.id).subscribe{
+                        view?.showMessage(R.string.details_delete_film)
+                        view?.showSaved(false)
+                    }
                 }
             },
-            {},
             {
-                val movie = Movie(
-                    movie.id,
-                    movie.title,
-                    movie.overview,
-                    movie.releaseData,
-                    movie.genres,
-                    movie.posterPath,
-                    movie.voteCount,
-                    movie.voteAverage
-                )
-                database.insertMovie(movie).subscribe {
-                    view?.showMessage(R.string.details_save_film)
-                    view?.showSaved(true)
-                }
+
             }
         )
     }
 
     override fun loadData(id: Long, lang: String) {
         view?.showEmptyProgress(true)
-        loadDetailsMovie(
-            id,
-            lang,
+        requestDisposable = network.fetchMovieDetails(id,lang).subscribe(
             {
+                view?.showData(it)
                 view?.showEmptyProgress(false)
             },
             {
@@ -71,47 +61,14 @@ class DetailsMoviePresenter(
 
     override fun refresh(id: Long, lang: String) {
         view?.showRefreshProgress(true)
-        loadDetailsMovie(
-            id,
-            lang,
+        requestDisposable = network.fetchMovieDetails(id,lang).subscribe(
             {
+                view?.showData(it)
                 view?.showRefreshProgress(false)
                 view?.showError(false)
             },
             {
                 view?.showRefreshProgress(false)
-            }
-        )
-    }
-
-    private fun loadDetailsMovie(id:Long,lang:String,onSuccess: () -> Unit, onError:(error:Throwable?)->Unit){
-        requestDisposable = network.getDetailFilm(id,lang)
-            .map(movieMapper)
-            .subscribe(
-                {
-                    checkExistenceMovieInTheDB(id,it,onSuccess)
-                },
-                {
-                    onError.invoke(it)
-                }
-            )
-    }
-
-    private fun checkExistenceMovieInTheDB(
-        id: Long,
-        detailsMovie: DetailsMovie,
-        onSuccess: () -> Unit
-    ) {
-        requestDisposable = database.existsFilm(id).subscribe(
-            {
-                detailsMovie.existsInTheDB = true
-                view?.showData(detailsMovie)
-                onSuccess.invoke()
-            },
-            {},
-            {
-                view?.showData(detailsMovie)
-                onSuccess.invoke()
             }
         )
     }
